Build libcmark locally as a build dependency
Prior to this commit, libcmark was expected to be installed by the user.

This commit migrates that step as a build-dependency of the Cargo package.

The native `libcmark` library is provided as a static library, thus the link attribute
references to the extern functions are removed.

The cmark project is referenced as a git submodule, at `9f8ef820301951f36301c1a40d036cafeaa78619`.