Remove the capability system and rework the resource management strategy
Previously, the library was architected according to a capability pattern.
This proved to be an unecessary layer of complication and indirection.
This commit re-architects the library into a simpler pattern where Nodes
are represented as enum variants that contain concrete Node types. Common
methods are implemented on the Node enum while methods specific to particular
node types are implemented on the concrete types.