Adjust mutability requirements of methods
Prior to this commit, none of the methods on Node or any of the
concrete Node types required a mutable self. This was misleading
since some of the methods actually do mutate the underlying data
structures. This commit changes the function signatures to reflect
whether or not something is actually getting mutated. It also
shrinks the scope of many of the unsafe blocks down to the bare
minimum.