Replace `as_resource()` with a NodeResource trait
Previously, to get at the underlying memory resources of a `Node`, there was
an `as_resource()` method that returned the `Resource` object for that `Node`s
data. Adding this extra transformation was unecessary and cumbersome to use.
This commit changes the approach to using a private trait which exposes the
fields directly.